% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_decimals_dataset.R
\name{change_decimals_dataset}
\alias{change_decimals_dataset}
\title{Takes a dataset and desired column and changes column to desired # of decimal places}
\usage{
change_decimals_dataset(data, col_name, n)
}
\arguments{
\item{data}{Dataset}

\item{col_name}{Name of column to be changed}

\item{n}{Desired number of digits AFTER decimal}
}
\value{
mutated dataset
}
\description{
Takes a dataset and desired column and changes column to desired # of decimal places
}
