% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mutate_groups.R
\name{data_mutate_groups}
\alias{data_mutate_groups}
\title{Takes a dataset and mutates it so that it returns one, at most, of a certain categorical
variable and the top value of another variable}
\usage{
data_mutate_groups(data, group_name, top_value_name)
}
\arguments{
\item{data}{The dataset}

\item{top_value_name}{The variable we want the top value from}

\item{groups}{The variable to group the data by}
}
\value{
The mutated dataset
}
\description{
Takes a dataset and mutates it so that it returns one, at most, of a certain categorical
variable and the top value of another variable
}
