% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/judge_team.R
\name{judge_team}
\alias{judge_team}
\title{Takes a football team name (presumably user's favorite) and determines how cool you seem}
\usage{
judge_team()
}
\value{
A judgement
}
\description{
Takes a football team name (presumably user's favorite) and determines how cool you seem
}
