% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_conversionsup.R
\name{length_conversionsup}
\alias{length_conversionsup}
\title{Takes a value and a US unit and converts value to different unit}
\usage{
length_conversionsup(value, measurement = "in", second_measure = "ft")
}
\arguments{
\item{value}{value associated with measurement}

\item{measurement}{The measurement of the value that is inputted}

\item{second_measure}{The measurement of the value that is to be outputted}
}
\value{
The value with a unit conversion
}
\description{
Takes a value and a US unit and converts value to different unit
}
