% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quotefromPatrick.R
\name{quotefromPatrick}
\alias{quotefromPatrick}
\title{Just gives back a random quote from patrick star}
\usage{
quotefromPatrick()
}
\value{
Quote
}
\description{
Just gives back a random quote from patrick star
}
