% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scientific_notation.R
\name{scientific_notation}
\alias{scientific_notation}
\title{Takes a dataset and desired column and changes column to desired # of decimal places}
\usage{
scientific_notation(nums, sig_figs)
}
\arguments{
\item{nums}{A vector of numbers to be translated to scientific notation}

\item{sig_figs}{desired Number of significant figures}
}
\value{
Number in scientific notation
}
\description{
Takes a dataset and desired column and changes column to desired # of decimal places
}
