% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance_level_concl.R
\name{significance_level_concl}
\alias{significance_level_concl}
\title{Similar to t.test, but considers significance levels}
\usage{
significance_level_concl(linear_model)
}
\arguments{
\item{linear_model}{The linear model of interest}

\item{fit_or_coef}{Whether we want p-value of linear fit or coefficient}
}
\value{
Tibble with analysis of p-value
}
\description{
Similar to t.test, but considers significance levels
}
